+++ python_lib/pywasm_ui/widgets/html/__init__.py
@@ -10,8 +10,10 @@ from .ContainerWidget import ContainerWidget
 from .DatePickerWidget import DatePickerWidget
 from .DividerWidget import DividerWidget
 from .HeadingWidget import HeadingWidget
+from .ImageWidget import ImageWidget
 from .IconButtonWidget import IconButtonWidget
 from .LabelWidget import LabelWidget
+from .LinkWidget import LinkWidget
 from .ListViewWidget import ListViewWidget
 from .ModalWidget import ModalWidget
 from .OptionWidget import OptionWidget
@@ -21,6 +23,7 @@ from .RowWidget import RowWidget
 from .SelectWidget import SelectWidget
 from .SliderWidget import SliderWidget
 from .StackWidget import StackWidget
+from .CodeBlockWidget import CodeBlockWidget
 from .TabItemWidget import TabItemWidget
 from .TabsWidget import TabsWidget
 from .TextAreaWidget import TextAreaWidget
@@ -40,10 +43,13 @@ __all__ = [
     "DatePickerWidget",
     "DividerWidget",
     "HeadingWidget",
+    "ImageWidget",
     "IconButtonWidget",
     "LabelWidget",
+    "LinkWidget",
     "ButtonWidget",
     "ParagraphWidget",
+    "CodeBlockWidget",
     "ModalWidget",
     "OptionWidget",
     "ProgressWidget",